+++ mobile/android/android-components/components/browser/session/src/main/java/mozilla/components/browser/session/engine/EngineObserver.kt
@@ -8,8 +8,8 @@ import android.graphics.Bitmap
 import android.os.Environment
 import mozilla.components.browser.session.Download
 import mozilla.components.browser.session.Session
+import mozilla.components.browser.session.engine.request.LoadRequestMetadata
 import mozilla.components.browser.session.engine.request.LoadRequestOption
-import mozilla.components.browser.session.engine.request.plus
 import mozilla.components.concept.engine.EngineSession
 import mozilla.components.concept.engine.HitResult
 import mozilla.components.concept.engine.manifest.WebAppManifest
@@ -48,20 +48,18 @@ internal class EngineObserver(
         }
     }
 
-    override fun onLoadRequest(triggeredByRedirect: Boolean, triggeredByWebContent: Boolean) {
+    override fun onLoadRequest(url: String, triggeredByRedirect: Boolean, triggeredByWebContent: Boolean) {
         if (triggeredByWebContent) {
             session.searchTerms = ""
         }
-        var triggers = LoadRequestOption.NONE.toMask()
-        if (triggeredByRedirect) {
-            triggers += LoadRequestOption.REDIRECT
-        }
 
-        if (triggeredByWebContent) {
-            triggers += LoadRequestOption.WEB_CONTENT
-        }
-
-        session.loadRequestTriggers = triggers
+        session.loadRequestMetadata = LoadRequestMetadata(
+            url,
+            arrayOf(
+                if (triggeredByRedirect) LoadRequestOption.REDIRECT else LoadRequestOption.NONE,
+                if (triggeredByWebContent) LoadRequestOption.WEB_CONTENT else LoadRequestOption.NONE
+            )
+        )
     }
 
     override fun onTitleChange(title: String) {