+++ mobile/android/android-components/components/browser/session/src/test/java/mozilla/components/browser/session/SessionTest.kt
@@ -10,8 +10,8 @@ import kotlinx.coroutines.GlobalScope
 import kotlinx.coroutines.async
 import kotlinx.coroutines.runBlocking
 import mozilla.components.browser.session.Session.Source
+import mozilla.components.browser.session.engine.request.LoadRequestMetadata
 import mozilla.components.browser.session.engine.request.LoadRequestOption
-import mozilla.components.browser.session.engine.request.isSet
 import mozilla.components.browser.session.tab.CustomTabConfig
 import mozilla.components.concept.engine.HitResult
 import mozilla.components.concept.engine.manifest.WebAppManifest
@@ -171,10 +171,13 @@ class SessionTest {
         val session = Session("https://www.mozilla.org")
         session.register(observer)
 
-        session.loadRequestTriggers = LoadRequestOption.REDIRECT.toMask()
+        session.loadRequestMetadata = LoadRequestMetadata(
+            "https://www.mozilla.org",
+            arrayOf(LoadRequestOption.REDIRECT)
+        )
 
-        assertTrue(session.loadRequestTriggers.isSet(LoadRequestOption.REDIRECT))
-        verify(observer, times(1)).onLoadRequest(eq(session), eq(true), eq(false))
+        assertTrue(session.loadRequestMetadata.isSet(LoadRequestOption.REDIRECT))
+        verify(observer, times(1)).onLoadRequest(eq(session), eq("https://www.mozilla.org"), eq(true), eq(false))
         verifyNoMoreInteractions(observer)
     }